add entity.py
