+++ homeassistant/components/ipma/sensor.py
@@ -14,8 +14,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.util import Throttle
 
-from . import IPMADevice
 from .const import DATA_API, DATA_LOCATION, DOMAIN, MIN_TIME_BETWEEN_UPDATES
+from .entity import IPMADevice
 
 _LOGGER = logging.getLogger(__name__)
